Introduce assignment 4D variable
This patch changes the code to consider a multiprocessor environment.
The task to processor/cluster to frequency is guided by means of a
binary variable assignment.

assignment is read from the input model. The schedulability test
analysis is done per processor.

The resources is expected to be split per processor. Currently
the code does not consider interprocessor communication. So, the
PCP applied is only per processor. Tasks are not allowed to share
resources between tasks in other processors.

The representation of task's priorities has changed. Task's priorities
are read from the task model. It simplifies the ordering assumption.

Signed-off-by: Eduardo Valentin <Eduardo.Valentin@icomp.ufam.edu.br>