Introduce GCD/LCM computation of tasks' period
The LCM of tasks' periods is commonly used on
system energy computation. The system energy is
usually computed during the LCM because the
tasks would then repeat their execution.

This implementation uses a hash table to cache
the most frequently computed LCMs. The hash table
contains offline computed LCMs of number from 0
to 100, which are the 80% of tasks with period between
40 and 70ms.

Signed-off-by: Eduardo Valentin <Eduardo.Valentin@icomp.ufam.edu.br>