solvers: exact: add a solver with response time analysis.
This change introduces a solver which computes tasks
response time. The computation is done by using
a Goal programming approach. The Goal is added
in the CPLEX solving procedure.

Therefore, in each node the .execute() callback will
be called. At this time, we check for the feasibility
of the current node. If the node is not feasible based
on response time analysis, we indicate to cplex to prune
the node. If the node is feasible, we do the typical
B&B procedure, by adding two subnodes based on the
real relaxation.

Signed-off-by: Eduardo Valentin <Eduardo.Valentin@icomp.ufam.edu.br>