hydra_solver: implement utilization based task model generation
This change adds the option to generate task models based on
utilization threshold. The strategy, instead of having a fixed
number of tasks, is to generate tasks until the requested
threshold is achieved. The tasks periods and utilizations are
random generated, the maximum execution time is then derived
based on the highest frequency.

The current limitations is how the cycles are generated for
different processors. As of the current implementation, each
processor will receive a multiplier, which will be used
to generate the amount of cycles based on the computed cycles
of the highest frequency.

Signed-off-by: Eduardo Valentin <Eduardo.Valentin@icomp.ufam.edu.br>