hydra_solvers: teach the simulator to read solvers commands from config
With this change, the simulator now knows how to read a configuration
file with the list of solvers to evaluate. The solvers config file
can be specified by command line parameter. The parameter is mandatory.

The config file format is:
<nsolvers> <nparams>
<solver 1 command line (including parameters, e.g. -tm)>
<solver 2 command line (including parameters, e.g. -tm)>
...
<solver nsolvers command line (including parameters, e.g. -tm)>

For now, the simulator will assume only 3 output parameters.
Each solver then is expected to output only three numbers
when done. Their meaning is:
<feasibility>
<execution time>
<total estimated energy>

Signed-off-by: Eduardo Valentin <Eduardo.Valentin@icomp.ufam.edu.br>