solvers: fix implementation of non-linear constraints
The linear constraints of utilization and response time
analysis were implemented using Goals in Concert/Cplex.
The Goal programming was not going through the possible
configurations and stopping at the very first iterations.

This patch fixes the problem by implementing the non-linear
constraints by means of IncumbentCallbackI. The callback
allows us to say if a node is feasible or not. We simply
call reject() whenever the decision variable is turns out
to be an infeasible configuration.

In this patch we also add the relaxed constraint of having
each processor at a maximum of 100% of capacity. The real
constraint is implemented in the IncumbentCallbackI.

One issue is that the presolver does not call the IncumbentCallbackI.
For this reason, to be able to implement the non-linear
constrains, the presolver is disable in both solvers.

Signed-off-by: Eduardo Valentin <Eduardo.Valentin@icomp.ufam.edu.br>