solvers: simplify usage of incumbent callback
On those solvers that use the incumbent callback
to implement nonlinear constraints, this patch
simplifies the implementation by removing the
temporary model file.

Now we pass the NumVar array directly and use
cplex to getValues. This way we do not need
to use a temporary file just to get the
Var array. Also, it simplifies the implementation
because it is not needed to linearize the matrix.

This patch should not change the functional behavior.
It is a simple optimization.

Signed-off-by: Eduardo Valentin <eduardo.valentin@icomp.ufam.edu.br>